spreadPosition double
git-svn-id: https://repo.algotrader.ch/svn/trunk/AlgoTrader@3398 19eef36c-86f0-de11-afab-0004ac968479

+++ src/com/algoTrader/service/ib/IbTransactionServiceImpl.java
@@ -236,7 +236,7 @@ public class IbTransactionServiceImpl extends IbTransactionServiceBase implement
             }
 
             PartialOrder partialOrder = order.getCurrentPartialOrder();
-            partialOrder.setSpreadPosition(Long.parseLong(spreadPosition));
+            partialOrder.setSpreadPosition(Double.parseDouble(spreadPosition));
 
             placeOrModifyPartialOrder(partialOrder, tick);
